+++ semantic_goals_generator/src/semantic_goals_generator.cpp
@@ -1,7 +1,7 @@
 /*
  * SEMANTIC GOALS GENERATOR ROS NODE
  *
- * Copyright (c) 2020-2021 Alberto José Tudela Roldán <ajtudela@gmail.com>
+ * Copyright (c) 2020-2022 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of semantic_navigation.
  * 
@@ -9,15 +9,19 @@
  *
  */
 
+// C++
+#include <limits>
+
+// ROS
 #include <tf/tf.h>
 #include <geometry_msgs/Point.h>
 #include <geometry_msgs/PoseArray.h>
 #include <geometry_msgs/PolygonStamped.h>
 #include <visualization_msgs/MarkerArray.h>
 #include <jsk_recognition_msgs/PolygonArray.h>
-
 #include <simple_laser_geometry/point2D.h>
 
+// Semantic Goals
 #include "semantic_goals_generator/semantic_goals_generator.h"
 
 /* Initialize the subscribers and publishers */
@@ -41,6 +45,7 @@ SemanticGoalsGenerator::SemanticGoalsGenerator(ros::NodeHandle& node, ros::NodeH
 SemanticGoalsGenerator::~SemanticGoalsGenerator(){
 	nodePrivate_.deleteParam("map_topic");
 	nodePrivate_.deleteParam("is_costmap");
+	nodePrivate_.deleteParam("full_map");
 }
 
 /* Update parameters of the node */
@@ -49,7 +54,12 @@ void SemanticGoalsGenerator::getParams(){
 
 	nodePrivate_.param<std::string>("map_topic", mapTopic_, "map");
 	nodePrivate_.param<bool>("is_costmap", isCostmap_, false);
+	nodePrivate_.param<bool>("full_map", fullMap_, false);
 	roiList_ = getROIParams();
+	if (roiList_.empty()){
+		ROS_ERROR("[Semantic goals generator]: The list of ROIs could not be found");
+		exit(1);
+	}
 }
 
 /* Map callback */
@@ -95,7 +105,8 @@ std::vector<ROI> SemanticGoalsGenerator::getROIParams(){
 			rois.push_back(roi);
 		}
 	}else{
-		ROS_ERROR("[Semantic goals generator]: Param 'rois' not exist");
+		ROS_ERROR("[Semantic goals generator]: Param 'rois' not exists");
+		return std::vector<ROI>();
 	}
 
 	ROS_INFO("[Semantic goals generator]: ROIs read");
@@ -103,29 +114,24 @@ std::vector<ROI> SemanticGoalsGenerator::getROIParams(){
 }
 
 /* Calculate bounding box for ROI */
-void SemanticGoalsGenerator::processBoundingBox(){
-	// Inflation radius must be positive
-	if(inflationRadius_ < 0) inflationRadius_ = 0.5;
-
-	inflatedFootprintSize_ = int(inflationRadius_ / resolution_) + 1;
-
+void SemanticGoalsGenerator::processBoundingBox(ROI roi){
 	// Region of interest (ROI) must lie inside the map boundaries
 	// If ROI is empty, the whole map is treated as ROI by default
-	if(roi_.empty()){
+	if(roi.empty()){
 		bBoxMinX_ = mapMinX_;
 		bBoxMaxX_ = mapMaxX_;
 		bBoxMinY_ = mapMinY_;
 		bBoxMaxY_ = mapMaxY_;
-		ROS_INFO("[Semantic goals generator]: No ROI specified, full map is used.");
+		ROS_INFO("[Semantic goals generator]: No ROI specified, full map is used");
 	}else{
 		// If the ROI is outside the map, adjust to map boundaries
 		// Determine bounding box of ROI
-		bBoxMinX_ = std::numeric_limits<int>::infinity();
-		bBoxMaxX_ = -std::numeric_limits<int>::infinity();
-		bBoxMinY_ = std::numeric_limits<int>::infinity();
-		bBoxMaxY_ = -std::numeric_limits<int>::infinity();
+		bBoxMinX_ = std::numeric_limits<double>::infinity();
+		bBoxMaxX_ = -std::numeric_limits<double>::infinity();
+		bBoxMinY_ = std::numeric_limits<double>::infinity();
+		bBoxMaxY_ = -std::numeric_limits<double>::infinity();
 
-		slg::Polygon polygonRoi = roi_.polygon;
+		slg::Polygon polygonRoi = roi.polygon;
 		for(int e = 0; e < polygonRoi.size(); e++){
 			slg::Point2D p = polygonRoi.getEdge(e).a;
 
@@ -157,19 +163,24 @@ void SemanticGoalsGenerator::processBoundingBox(){
 bool SemanticGoalsGenerator::SemanticGoalsService(semantic_goals_generator::SemanticGoals::Request& req, semantic_goals_generator::SemanticGoals::Response& res){
 	ROS_INFO("[Semantic goals generator]: Incoming service request: %i, %s", req.n, req.roi_name.c_str());
 
-	// Clear previous ROI
-	roi_.clear();
+	ROI currentRoi;
 
 	// Get arguments
 	int n = req.n;
 	for(ROI roi: roiList_){
 		if(roi.polygon.getName() == req.roi_name){
-			roi_ = roi;
+			currentRoi = roi;
 			break;
 		}
 	}
 	direction_ = req.direction;
 	border_ = req.border;
+
+	// If the requested ROI is empty and we don't want to use the full map
+	if (currentRoi.empty() && !fullMap_){
+		ROS_FATAL("[Semantic goals generator]: The requested ROI: %s, could not be found in the list", req.roi_name.c_str());
+		return false;
+	}
 
 	// Wait for map
 	nav_msgs::OccupancyGrid::ConstPtr msgMap = ros::topic::waitForMessage<nav_msgs::OccupancyGrid>(mapTopic_, node_, ros::Duration(10));
@@ -180,8 +191,12 @@ bool SemanticGoalsGenerator::SemanticGoalsService(semantic_goals_generator::Sema
 		return false;
 	}
 
+	// Inflation radius must be positive
+	if(inflationRadius_ < 0) inflationRadius_ = 0.5;
+	inflatedFootprintSize_ = int(inflationRadius_ / resolution_) + 1;
+
 	// Process bounding box
-	processBoundingBox();
+	processBoundingBox(currentRoi);
 
 	// Generate response
 	res.goals.header.frame_id = mapTopic_;
@@ -194,8 +209,6 @@ bool SemanticGoalsGenerator::SemanticGoalsService(semantic_goals_generator::Sema
 	std::uniform_real_distribution<double> distPI(0.0, 2 * M_PI);
 
 	int count = 0;
-	//int upperBound = n  * ( 2 +  inflationRadius_ / 0.01);
-	//while( (res.goals.poses.size() < n) && (count < upperBound) ){
 	while( (res.goals.poses.size() < n)){
 		count += 1;
 		int cellX = distX(gen);
@@ -206,16 +219,16 @@ bool SemanticGoalsGenerator::SemanticGoalsService(semantic_goals_generator::Sema
 		pose.position.y = cellY * resolution_ + origin_.position.y;
 
 		// If the point lies within ROI and is not in collision
-		if(roi_.inROI(pose.position.x, pose.position.y) && !inCollision(cellX, cellY) && roi_.disFromBorders(pose.position.x, pose.position.y, border_)){
+		if(currentRoi.inROI(pose.position.x, pose.position.y) && !inCollision(cellX, cellY) && currentRoi.disFromBorders(pose.position.x, pose.position.y, border_)){
 			// Generate orientation
 			double yaw;
 			if (direction_ == semantic_goals_generator::SemanticGoalsRequest::OUTSIDE){
-				yaw = atan2((pose.position.y - roi_.polygon.centroid().y), (pose.position.x - roi_.polygon.centroid().x));
+				yaw = atan2((pose.position.y - currentRoi.polygon.centroid().y), (pose.position.x - currentRoi.polygon.centroid().x));
 			}else if (direction_ == semantic_goals_generator::SemanticGoalsRequest::INSIDE){
-				yaw = atan2((pose.position.y - roi_.polygon.centroid().y), (pose.position.x - roi_.polygon.centroid().x)) + M_PI;
+				yaw = atan2((pose.position.y - currentRoi.polygon.centroid().y), (pose.position.x - currentRoi.polygon.centroid().x)) + M_PI;
 			}else if (direction_ == semantic_goals_generator::SemanticGoalsRequest::STORED){
-				if (roi_.yaw > -M_PI && roi_.yaw < M_PI){
-					yaw = roi_.yaw;
+				if (currentRoi.yaw > -M_PI && currentRoi.yaw < M_PI){
+					yaw = currentRoi.yaw;
 				}else{
 					yaw = distPI(gen);
 				}
@@ -229,7 +242,7 @@ bool SemanticGoalsGenerator::SemanticGoalsService(semantic_goals_generator::Sema
 				yaw = distPI(gen);
 			}
 			tf::quaternionTFToMsg(tf::createQuaternionFromYaw(yaw), pose.orientation);
-			ROS_INFO("[Semantic goals generator]: Pose (x: %f, y: %f, yaw: %f)", pose.position.x, pose.position.y, yaw);
+			ROS_INFO("[Semantic goals generator]: Pose %i (x: %f, y: %f, yaw: %f)", res.goals.poses.size()+1, pose.position.x, pose.position.y, yaw);
 
 			res.goals.poses.push_back(pose);
 		}